requirements.txt
Pillow
stegano

